requirements.txt
absl-py==0.13.0
gym==0.18.3
numpy==1.21.1
Pillow==8.2.0
pygame==2.0.1
pyglet==1.5.15
pymunk==5.6.0